+++ src/web_server/__main__.py
@@ -1,0 +1,7 @@
+"""Entry point for the module"""
+# Local Imports
+from .server.utils import run_server
+
+
+if __name__ == "__main__":
+    run_server()
