+++ src/web_server/utils.py
@@ -1,6 +1,9 @@
+# Default Imports
 import subprocess
 import sys
 
+# Local Imports
+from . import constants as CONSTANTS
 
 def run_echo_new_window(message: str):
     if sys.platform == "win32":
@@ -22,3 +25,5 @@ def run_echo_new_window(message: str):
             )
         except FileNotFoundError:
             subprocess.Popen(["x-terminal-emulator", "-e", f"echo {message}; bash"])
+
+
